Add mapper

+++ src/main/java/dev/chat/service/ChatService.java
@@ -5,13 +5,13 @@ import dev.chat.entity.Chat;
 import dev.chat.entity.User;
 import dev.chat.handler.ChatNotFoundException;
 import dev.chat.handler.UserNotFoundException;
+import dev.chat.mapper.ChatMapper;
 import dev.chat.repository.ChatRepository;
 import dev.chat.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -19,14 +19,15 @@ public class ChatService {
 
     private final ChatRepository chatRepository;
     private final UserRepository userRepository;
+    private final ChatMapper chatMapper;
 
     @Autowired
-    public ChatService(ChatRepository chatRepository, UserRepository userRepository) {
+    public ChatService(ChatRepository chatRepository, UserRepository userRepository, ChatMapper chatMapper) {
         this.chatRepository = chatRepository;
         this.userRepository = userRepository;
+        this.chatMapper = chatMapper;
     }
 
-    // Создать чат
     public ChatDto createChat(String chatName, List<Long> participantIds) {
         List<User> participants = userRepository.findAllById(participantIds);
         Chat chat = Chat.builder()
@@ -34,39 +35,31 @@ public class ChatService {
                 .participants(participants)
                 .build();
         Chat savedChat = chatRepository.save(chat);
-        return convertToChatDto(savedChat);
+        return chatMapper.chatToChatDTO(savedChat);
     }
-
-    // Получить все чаты для пользователя
-    //todo add pagination
+    //todo добавить пагинацию
     public List<ChatDto> getAllChatsForUser(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new UserNotFoundException("User with ID " + userId + " not found"));
-        List<Chat> chats = chatRepository.findAllByParticipantsContains(user);
-        return chats.stream().map(this::convertToChatDto).collect(Collectors.toList());
+        User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException("User not found with ID: " + userId));
+        List<Chat> chats = chatRepository.findAllByParticipantsContaining(user);
+        return chats.stream().map(chatMapper::chatToChatDTO).collect(Collectors.toList());
     }
 
-    // Переименовать чат
+    public ChatDto getChatById(Long chatId) {
+        Chat chat = chatRepository.findById(chatId).orElseThrow(() -> new ChatNotFoundException("Chat not found with ID: " + chatId));
+        return chatMapper.chatToChatDTO(chat);
+    }
+
     public ChatDto renameChat(Long chatId, String newChatName) {
-        //add controller advice , handle
-        Chat chat = chatRepository.findById(chatId)
-                .orElseThrow(() -> new ChatNotFoundException("Chat with ID " + chatId + " not found"));
+        Chat chat = chatRepository.findById(chatId).orElseThrow(() -> new ChatNotFoundException("Chat not found with ID: " + chatId));
         chat.setChatName(newChatName);
         Chat savedChat = chatRepository.save(chat);
-        return convertToChatDto(savedChat);
+        return chatMapper.chatToChatDTO(savedChat);
     }
 
-    // Удалить чат
     public void deleteChat(Long chatId) {
+        if (!chatRepository.existsById(chatId)) {
+            throw new ChatNotFoundException("Chat not found with ID: " + chatId);
+        }
         chatRepository.deleteById(chatId);
-    }
-
-    // Преобразовать сущность Chat в DTO
-    private ChatDto convertToChatDto(Chat chat) {
-        //todo mapstruct
-        ChatDto chatDto = new ChatDto();
-        chatDto.setChatID(chat.getChatId());
-        chatDto.setChatName(chat.getChatName());
-        return chatDto;
     }
 }
